Derive and persist node metadata when adding or editing nodes
